requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
qrcode==7.4.2
python-decouple==3.8
psycopg2-binary==2.9.9
django-ratelimit==4.1.0
geoip2==4.7.0
user-agents==2.2.0
django-extensions==3.2.3
openpyxl==3.1.2
